+++ src/app/robots.ts
@@ -13,7 +13,6 @@ export default function robots(): MetadataRoute.Robots {
           "/admin/",
           "/admin/*",
           "/api/auth/*",
-          "/vouchery/overeni",
           "/rezervace/storno/*",
           "/rezervace/sprava/*",
           "/rezervace/akce/*",